Update RU_Глава_01.md
A few edits. More editorial in style than anything else. A lot of paraphrasing (hopefully) enabling easier thought transitions. 

The original translator should probably be looking all of this over for final approval, as a lot of it is subjective. Things that sound good to me may not sound good to the final target audience.